add code for Get-PSResourceRepository and move to srcNew
