+++ src/code/RegisterPSResourceRepository.cs
@@ -1,35 +1,33 @@
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
 using System;
 using System.Collections;
 using System.Management.Automation;
 using System.Globalization;
 using Microsoft.PowerShell.PowerShellGet.RepositorySettings;
-using Microsoft.PowerShell.PowerShellGet.PSRepositoryItem;
 using System.Collections.Generic;
 using System.Linq;
+using System.IO;
 
 namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
 {
     /// <summary>
-    /// todo: fill
+    /// The Register-PSResourceRepository cmdlet replaces the Register-PSRepository from V2.
+    /// It registers a repository for PowerShell modules. The repository is registered to the current user's scope and does not have a system-wide scope.
     /// </summary>
 
-    [Cmdlet(VerbsLifecycle.Register,
-        "PSResourceRepository",
-        DefaultParameterSetName = "NameParameterSet",
-        SupportsShouldProcess = true,
-        HelpUri = "<add>",
-        RemotingCapability = RemotingCapability.None)]
+    [Cmdlet(VerbsLifecycle.Register, "PSResourceRepository", DefaultParameterSetName = "NameParameterSet", SupportsShouldProcess = true,
+        HelpUri = "<add>", RemotingCapability = RemotingCapability.None)]
     public sealed
-    class NEWRegisterPSResourceRepository : PSCmdlet
+    class RegisterPSResourceRepository : PSCmdlet
     {
         private string PSGalleryRepoName = "PSGallery";
         private string PSGalleryRepoURL = "https://www.powershellgallery.com/api/v2";
-        private const int defaultPriority = 50;
-        private const bool defaultTrusted = false;
 
-        #region Parameters
         /// <summary>
-        /// Specifies name for the repository to be registered.
+        /// Specifies the desired name for the repository to be registered.
         /// </summary>
         [Parameter(Mandatory = true, Position = 0, ParameterSetName = "NameParameterSet")]
         [ValidateNotNullOrEmpty]
@@ -54,12 +52,28 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
             { return _url; }
 
             set
-            { _url = value; } //todo: actually set code here
+            {
+                Uri url;
+                if (!Uri.TryCreate(value, string.Empty, out url))
+                {
+                    // Try the URL as a file path
+                    var resolvedPath = string.Format(CultureInfo.InvariantCulture, "{0}{1}{2}", Uri.UriSchemeFile, Uri.SchemeDelimiter, SessionState.Path.GetResolvedPSPathFromPSPath(value.ToString()).FirstOrDefault().Path);
+                    if (!Uri.TryCreate(resolvedPath, UriKind.Absolute, out url))
+                    {
+                        var message = string.Format(CultureInfo.InvariantCulture, "The URL provided is not valid: {0}", value);
+                        var ex = new ArgumentException(message);
+                        var moduleManifestNotFound = new ErrorRecord(ex, "InvalidUrl", ErrorCategory.InvalidArgument, null);
+                        ThrowTerminatingError(moduleManifestNotFound);
+                    }
+                }
+
+                _url = url;
+            }
         }
         private Uri _url;
 
         /// <summary>
-        /// When specified, registers PSGallery repository.
+        /// Registers the PowerShell Gallery.
         /// </summary>
         [Parameter(Mandatory = true, ParameterSetName = "PSGalleryParameterSet")]
         public SwitchParameter PSGallery
@@ -73,54 +87,30 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         private SwitchParameter _psgallery;
 
         /// <summary>
-        /// Specifies a hashtable of repositories and is used to register multiple repositories at once.
+        /// Repositories is a hashtable and is used to register multiple repositories at once.
         /// </summary>
         [Parameter(Mandatory = true, ParameterSetName = "RepositoriesParameterSet")]
         [ValidateNotNullOrEmpty]
         public List<Hashtable> Repositories
         {
-            get
-            { return _repositories; }
+            get { return _repositories; }
 
-            set
-            { _repositories = value; }
+            set { _repositories = value; }
         }
         private List<Hashtable> _repositories;
 
         /// <summary>
         /// Specifies whether the repository should be trusted.
         /// </summary>
-        [Parameter(ParameterSetName = "NameParameterSet")]
         [Parameter(ParameterSetName = "PSGalleryParameterSet")]
+        [Parameter(ParameterSetName = "NameParameterSet")]
         public SwitchParameter Trusted
         {
-            get
-            { return _trusted; }
+            get { return _trusted; }
 
-            set
-            { _trusted = value; }
+            set { _trusted = value; }
         }
         private SwitchParameter _trusted;
-
-        /// <summary>
-        /// Specifies the priority ranking of the repository, such that repositories with higher ranking priority are searched
-        /// before a lower ranking priority one, when searching for a repository item across multiple registered repositories.
-        /// Valid priority values range from 0 to 50, such that a lower numeric value (i.e 10) corresponds
-        /// to a higher priority ranking than a higher numeric value (i.e 40). Has default value of 50.
-        /// </summary>
-        [Parameter(ParameterSetName = "NameParameterSet")]
-        [Parameter(ParameterSetName = "PSGalleryParameterSet")]
-        [ValidateNotNullOrEmpty]
-        [ValidateRange(0, 50)]
-        public int Priority
-        {
-            get
-            { return _priority; }
-
-            set
-            { _priority = value; }
-        }
-        private int _priority = defaultPriority;
 
         /// <summary>
         /// Specifies a proxy server for the request, rather than a direct connection to the internet resource.
@@ -144,123 +134,127 @@ namespace Microsoft.PowerShell.PowerShellGet.Cmdlets
         public PSCredential ProxyCredential
         {
             get
-            { return _proxycredential; }
+            { return _proxyCredential; }
 
             set
-            { _proxycredential = value; }
+            { _proxyCredential = value; }
         }
-        private PSCredential _proxycredential;
+        private PSCredential _proxyCredential;
 
         /// <summary>
-        /// When specified, displays the succcessfully registered repository and its information
+        /// The following is the definition of the input parameter "Port".
+        /// Specifies the port to be used when connecting to the ws management service.
         /// </summary>
-        [Parameter]
-        public SwitchParameter PassThru
+        [Parameter(ParameterSetName = "PSGalleryParameterSet")]
+        [Parameter(ParameterSetName = "NameParameterSet")]
+        [ValidateNotNullOrEmpty]
+        [ValidateRange(0, 50)]
+        public int Priority
         {
-            get
-            { return _passThru; }
+            get { return _priority; }
 
-            set
-            { _passThru = value; }
+            set { _priority = value; }
         }
-        private SwitchParameter _passThru;
-        #endregion
+        private int _priority = 50;
 
+        /// <summary>
+        /// </summary>
         protected override void ProcessRecord()
         {
-            WriteDebug("new cmdlet");
             var r = new RespositorySettings();
-            List<PSRespositoryItem> items = new List<PSRespositoryItem>();
 
-            if(ParameterSetName.Equals("PSGalleryParameterSet")){
-                PSGalleryParameterSetHelper(r, items);
-            }
-            else if(ParameterSetName.Equals("NameParameterSet"))
+            if (ParameterSetName.Equals("PSGalleryParameterSet"))
             {
-                NameParameterSetHelper(r, items);
-            }
-            else if(ParameterSetName.Equals("RepositoriesParameterSet"))
-            {
-                RepositoriesParameterSetHelper(r, items);
-            }
-
-            if(_passThru)
-            {
-                foreach (var item in items)
+                if (!_psgallery)
                 {
-                    WriteObject(item);
+                    return;
                 }
-            }
 
-        }
+                /// collect parameters and make one call
+                var psGalleryUri = new Uri(PSGalleryRepoURL);
 
-        private void PSGalleryParameterSetHelper(RespositorySettings rs, List<PSRespositoryItem> items)
-        {
-            var psGalleryUri = new Uri(PSGalleryRepoURL);
-            try
-            {
-                items.Add(rs.Add(PSGalleryRepoName, psGalleryUri, _priority, _trusted));
-            }
-            catch (Exception e)
-            {
-                throw new Exception(e.Message);
-            }
-        }
-
-        private void NameParameterSetHelper(RespositorySettings rs, List<PSRespositoryItem> items)
-        {
-            if(_name.Equals("PSGallery"))
-            {
-                throw new ArgumentException(String.Format("Use 'Register-PSResourceRepository -PSGallery' to register the PSGallery repository."));
-            }
-            else{
                 try
                 {
-                    items.Add(rs.Add(_name, _url, _priority, _trusted));
+                    r.Add(PSGalleryRepoName, psGalleryUri, _priority, _trusted);
                 }
-                catch(Exception e)
+                catch (Exception e)
                 {
                     throw new Exception(e.Message);
                 }
+
             }
-
-        }
-
-        private void RepositoriesParameterSetHelper(RespositorySettings rs, List<PSRespositoryItem> items)
-        {
-            foreach(var repo in _repositories)
+            else if (ParameterSetName.Equals("NameParameterSet"))
             {
-                if(repo.ContainsKey(PSGalleryRepoName))
-                {
-                    _priority = repo.ContainsKey("Priority") ? (int)repo["Priority"] : defaultPriority;
-                    _trusted = repo.ContainsKey("Trusted") ? (bool)repo["Trusted"] : defaultTrusted;
-                    PSGalleryParameterSetHelper(rs, items);
-                    continue;
-                }
-
-                // if not PSGallery, then check that sufficient keys exist for NameParameterSet style register option
-                if(!repo.ContainsKey("Name") || String.IsNullOrEmpty(repo["Name"].ToString()))
+                if (String.IsNullOrEmpty(_name))
                 {
                     throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Repository name cannot be null"));
                 }
-                if(!repo.ContainsKey("Url") || String.IsNullOrEmpty(repo["Url"].ToString()))
+                if (String.IsNullOrEmpty(_url.ToString()))
                 {
                     throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Repository url cannot be null"));
                 }
 
-                Uri _repoURL;
-                if(!(Uri.TryCreate(repo["URL"].ToString(), UriKind.Absolute, out _repoURL)
-                    && (_repoURL.Scheme == Uri.UriSchemeHttp || _repoURL.Scheme == Uri.UriSchemeHttps || _repoURL.Scheme == Uri.UriSchemeFtp || _repoURL.Scheme == Uri.UriSchemeFile)))
+                try
                 {
-                        throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Invalid Url"));
+                    r.Add(_name, _url, _priority, _trusted);
                 }
+                catch (Exception e)
+                {
+                    throw new Exception(e.Message);
+                }
+            }
+            else if (ParameterSetName.Equals("RepositoriesParameterSet"))
+            {
+                foreach (var repo in _repositories)
+                {
+                    if (repo.ContainsKey(PSGalleryRepoName)   )
+                    {
+                        var _psGalleryRepoURL = new Uri(PSGalleryRepoURL);
+                        int _psGalleryRepoPriority = repo.ContainsKey("Priority") ? (int)repo["Priority"] : 50;
+                        var _psGalleryRepoTrusted = repo.ContainsKey("Trusted") ? (bool)repo["Trusted"] : false;
 
-                _name = repo["Name"].ToString();
-                _url = _repoURL;
-                _priority = repo.ContainsKey("Priority") ? Convert.ToInt32(repo["Priority"].ToString()) : defaultPriority;
-                _trusted = repo.ContainsKey("Trusted") ? Convert.ToBoolean(repo["Trusted"].ToString()) : defaultTrusted;
+                        r.Add(PSGalleryRepoName, _psGalleryRepoURL, _psGalleryRepoPriority, _psGalleryRepoTrusted);
+                        continue;
+                    }
 
-                NameParameterSetHelper(rs, items);
+                    // check if key exists
+                    if (!repo.ContainsKey("Name") || String.IsNullOrEmpty(repo["Name"].ToString()))
+                    {
+                        throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Repository name cannot be null"));
+                    }
+                    if (!repo.ContainsKey("Url") || String.IsNullOrEmpty(repo["Url"].ToString()))
+                    {
+                        throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Repository url cannot be null"));
+                    }
+
+                    // https://docs.microsoft.com/en-us/dotnet/api/system.uri.trycreate?view=netframework-4.8#System_Uri_TryCreate_System_Uri_System_Uri_System_Uri__
+                    // convert the string to a url and check to see if the url is formatted correctly
+                    Uri _repoURL;
+                    if (!(Uri.TryCreate(repo["URL"].ToString(), UriKind.Absolute, out _repoURL)
+                         && (_repoURL.Scheme == Uri.UriSchemeHttp || _repoURL.Scheme == Uri.UriSchemeHttps || _repoURL.Scheme == Uri.UriSchemeFtp || _repoURL.Scheme == Uri.UriSchemeFile)))
+                    {
+                        throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Invalid Url"));
+                    }
+
+                    int _repoPriority = 50;
+                    if (repo.ContainsKey("Priority"))
+                    {
+                        _repoPriority = Convert.ToInt32(repo["Priority"].ToString());
+                    }
+
+                    bool _repoTrusted = false;
+                    if (repo.ContainsKey("Trusted"))
+                    {
+                        _repoTrusted = Convert.ToBoolean(repo["Trusted"].ToString());
+                    }
+
+                    r.Add(repo["Name"].ToString(), _repoURL, _repoPriority, _repoTrusted);
+                }
+            }
+            else if (_name.Equals(PSGalleryRepoName))
+            {
+                //throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, Messages.UsePSGalleryParameterSetOnRegister));
+                throw new System.ArgumentException(string.Format(CultureInfo.InvariantCulture, "Use PSGallery Parmenter Set on Register"));
             }
         }
     }
